requirements.txt
fastapi
uvicorn
jinja2
python-multipart
python-dateutil
feedgen
email-validator
phonenumbers
slowapi
passlib
bcrypt
resend
python-dotenv
openai
python-dotenv
requests
feedparser
beautifulsoup4
newspaper3k
lxml_html_clean
boto3
